+++ CHANGELOG.md
@@ -6,9 +6,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.2.61] - 2020-11-07
+## [0.2.65] - 2020-11-08
+### Changed
+- Views.PageChangelog
+- Views.PageHttpClient
+- Views.PagePing
+- Views.PageProxy
+- Views.PageWebClient
+- Views.PageWebRequest
+- Views.WindowMain
+
+## [0.2.62] - 2020-11-07
 ### Changed
 - Moved ProxyEntity to nuget package Net.Utils (https://www.nuget.org/packages/Net.Utils/)
+- Moved HttpServiceEntity to nuget package Net.Utils (https://www.nuget.org/packages/Net.Utils/)
 
 ## [0.2.60] - 2020-11-05
 ### Added